script: Move reporting interfaces
Also moves some relevant methods from GlobalScope that
were only used within ReportingObserver.

Part of #38901
