+++ components/script/dom/window.rs
@@ -168,8 +168,8 @@ use crate::dom::navigator::Navigator;
 use crate::dom::node::{Node, NodeDamage, NodeTraits, from_untrusted_node_address};
 use crate::dom::performance::performance::Performance;
 use crate::dom::promise::Promise;
-use crate::dom::reportingendpoint::{ReportingEndpoint, SendReportsToEndpoints};
-use crate::dom::reportingobserver::ReportingObserver;
+use crate::dom::reporting::reportingendpoint::{ReportingEndpoint, SendReportsToEndpoints};
+use crate::dom::reporting::reportingobserver::ReportingObserver;
 use crate::dom::screen::Screen;
 use crate::dom::scrolling_box::{ScrollingBox, ScrollingBoxSource};
 use crate::dom::selection::Selection;
