+++ components/script/dom/workers/workerglobalscope.rs
@@ -66,8 +66,8 @@ use crate::dom::idbfactory::IDBFactory;
 use crate::dom::performance::performance::Performance;
 use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::promise::Promise;
-use crate::dom::reportingendpoint::{ReportingEndpoint, SendReportsToEndpoints};
-use crate::dom::reportingobserver::ReportingObserver;
+use crate::dom::reporting::reportingendpoint::{ReportingEndpoint, SendReportsToEndpoints};
+use crate::dom::reporting::reportingobserver::ReportingObserver;
 use crate::dom::trustedtypes::trustedscripturl::TrustedScriptURL;
 use crate::dom::trustedtypes::trustedtypepolicyfactory::TrustedTypePolicyFactory;
 use crate::dom::types::ImageBitmap;
